azure bd works

+++ src/main/java/com/homebudget/homeentirety/models/MonthlyExpenses.java
@@ -1,28 +1,41 @@
 package com.homebudget.homeentirety.models;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import com.sun.xml.bind.v2.model.core.ID;
+
+import javax.persistence.*;
+import java.util.List;
 
 @Entity(name = "expenses")
 public class MonthlyExpenses {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int ID;
+    private Long ID;
     private String Objective;
     private float Totally;
 
+//    @OneToMany(mappedBy = "expenses")
+//    @JoinTable(
+//            name = "expenses",
+//            joinColumns = {@JoinColumn(name = "Totally")},
+//            inverseJoinColumns = {@JoinColumn(name = "ObjectiveId")})
+//    private List<TradeListRepository> tradeList;
 
     public MonthlyExpenses() {
     }
 
+//    public List<TradeListRepository> getTradeList() {
+//        return tradeList;
+//    }
+//
+//    public void setTradeList(List<TradeListRepository> tradeList) {
+//        this.tradeList = tradeList;
+//    }
 
-    public int getID() {
+    public Long getID() {
         return ID;
     }
 
-    public void setID(int ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
